refactor: each modal on its own component
