+++ src/components/Reown/modals/SignOracleDataModal.js
@@ -1,0 +1,30 @@
+import React from 'react';
+import { t } from 'ttag';
+
+export function SignOracleDataModal({ data, onAccept, onReject }) {
+  return (
+    <>
+      <div className="modal-header">
+        <h5 className="modal-title">{t`Sign Oracle Data`}</h5>
+        <button type="button" className="close" data-dismiss="modal" aria-label="Close">
+          <span aria-hidden="true">&times;</span>
+        </button>
+      </div>
+      <div className="modal-body">
+        <div className="d-flex align-items-center mb-3">
+          {data.dapp.icon && <img src={data.dapp.icon} alt="dApp icon" className="mr-3" style={{ width: 48, height: 48 }} />}
+          <div>
+            <h6 className="mb-1">{data.dapp.proposer}</h6>
+            <small className="text-muted">{data.dapp.url}</small>
+          </div>
+        </div>
+        <p>{t`Oracle data to sign:`}</p>
+        <pre className="bg-light p-3 rounded">{JSON.stringify(data.data, null, 2)}</pre>
+      </div>
+      <div className="modal-footer">
+        <button type="button" className="btn btn-secondary" onClick={onReject} data-dismiss="modal">{t`Reject`}</button>
+        <button type="button" className="btn btn-hathor" onClick={onAccept}>{t`Sign`}</button>
+      </div>
+    </>
+  );
+}
